\name{caGeneFilter-package}
\Rdversion{1.1}
\alias{caGeneFilter-package}

\docType{package}

\title{
RWebServices interface to the genefilter package.
}

\description{
Provides an RWebServices interface to a part of the functionality provided
by the genefilter package.
}

\details{
\tabular{ll}{
Package: \tab caGeneFilter\cr
Type: \tab Package\cr
Version: \tab 0.1.17\cr
Date: \tab 2009-01-12\cr
License: \tab caBIG\cr
LazyLoad: \tab yes\cr
}
An overview of how to use the package, including the most important
functions
}

\author{
Tom Milac <tmilac@fhcrc.org>
}

%\references{
%Literature or other references for background information
%}

\keyword{ package }

%\seealso{
%Optional links to other man pages, e.g.
%\code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}

%\examples{
%simple examples of the most important functions
%}
