\name{Filter-class}
\docType{class}
\alias{Filter-class}

\title{Class "Filter"}
\description{
  Filter is the base class for all classes which encapsulate
  the parameters needed to run a gene filter analysis.
}

\section{Objects from the Class}{

  Objects can be created by calls to the constructor
  \code{\link{Filter}}.

}

%\section{Slots}{
%  \describe{
%  }
%}
\section{Methods}{

  See \code{\link{Filter}}

}

\author{Tom Milac \url{tmilac@fhcrc.org}}

\seealso{

  \code{\link{Filter}} for object construction, accessors

}

\examples{
showClass("Filter")
Filter()
}
\keyword{classes}
