\name{Filter}
\alias{Filter}

\title{Filter constructor and accessors}

\description{
  Functions for creating and manipulating Filter-class objects.
}

\usage{
Filter()
}

%\arguments{
%}

\value{

  \item{Filter}{Object of class
    \code{\linkS4class{Filter}}.}

}

\author{Tom Milac \url{tmilac@fhcrc.org}}

\seealso{\code{\linkS4class{Filter}-class}}

\examples{
obj <- Filter()
}

\keyword{manip}
